+++ classes/local/hook/output/before_standard_head_html_generation.php
@@ -14,20 +14,28 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
+namespace theme_boost_union\local\hook\output;
+
 /**
- * Theme Boost Union - Version file
+ * Theme Boost Union - Hook: Allows plugins to add any elements to the page <head> html tag.
  *
  * @package    theme_boost_union
- * @copyright  2022 Alexander Bias, lern.link GmbH <alexander.bias@lernlink.de>
+ * @copyright  2024 Alexander Bias <bias@alexanderbias.de>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+class before_standard_head_html_generation {
+    /**
+     * Callback to add head elements.
+     *
+     * @param \core\hook\output\before_standard_head_html_generation $hook
+     */
+    public static function callback(\core\hook\output\before_standard_head_html_generation $hook): void {
+        global $CFG;
 
-defined('MOODLE_INTERNAL') || die();
+        // Require local library.
+        require_once($CFG->dirroot.'/theme/boost_union/locallib.php');
 
-$plugin->component = 'theme_boost_union';
-$plugin->version = 2023102032;
-$plugin->release = 'v4.3-r10';
-$plugin->requires = 2023100900;
-$plugin->supported = [403, 403];
-$plugin->maturity = MATURITY_STABLE;
-$plugin->dependencies = ['theme_boost' => 2023100900];
+        // Call callback implementation.
+        theme_boost_union_callbackimpl_before_standard_html($hook);
+    }
+}